Feat: board_hashtag entity, hashtag entity 구현
