--- friends_server/src/main/kotlin/com/friends/board/repository/BoardRepository.kt
@@ -1,9 +1,0 @@
-package com.friends.board.repository
-
-import com.friends.board.entity.BoardEntity
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface BoardRepository: JpaRepository<BoardEntity, Long> {
-}
