--- friends_server/src/main/kotlin/com/friends/board/service/BoardService.kt
@@ -1,44 +1,0 @@
-package com.friends.board.service
-
-import com.friends.board.dto.BoardFormDto
-import com.friends.board.entity.BoardEntity
-import com.friends.board.repository.BoardRepository
-import com.friends.member.repository.MemberRepository
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.stereotype.Service
-
-@Service
-class BoardService @Autowired constructor(
-    val boardRepository: BoardRepository,
-    val memberRepository: MemberRepository
-){
-    fun save(boardFormDto: BoardFormDto): Long?{
-        val member = memberRepository.findById(boardFormDto.memberId)
-            .orElseThrow { IllegalArgumentException("Member not found with id: ${boardFormDto.memberId}")}
-
-        val board = BoardEntity(
-            member = member,
-            content = boardFormDto.content,
-        )
-
-        return boardRepository.save(board).id
-    }
-
-    fun getBoard(id: Long): BoardEntity {
-        return boardRepository.findById(id).get()
-    }
-
-    fun deleteBoard(id: Long){
-        return boardRepository.deleteById(id)
-    }
-
-    fun updateBoard(id: Long, boardFormDto: BoardFormDto): BoardEntity {
-        val board = boardRepository.findById(id).get()
-        board.updateBoard(boardFormDto)
-        return board
-    }
-
-    fun getBoardList(): List<BoardEntity>{
-        return boardRepository.findAll()
-    }
-}
